+++ examples/blog-studio/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @examples/hello-world-i18n-studio
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [51c2917]
+  - @tinloof/sanity-studio@1.16.2
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [17e5837]
+  - @tinloof/sanity-document-options@1.3.2
+  - @tinloof/sanity-studio@1.16.1
+
 ## 0.0.5
 
 ### Patch Changes