+++ packages/sanity-studio/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @tinloof/sanity-plugin-pages-navigator
 
+## 1.16.2
+
+### Patch Changes
+
+- 51c2917: Highlight decorator
+
+## 1.16.1
+
+### Patch Changes
+
+- Updated dependencies [c0fce45]
+  - @tinloof/sanity-web@1.1.1
+
 ## 1.16.0
 
 ### Minor Changes